commit fix userAPI<encoder>
