+++ TuneBox/Customer/src/main/java/org/example/customer/API/ChangePassword.java
@@ -13,15 +13,5 @@ public class ChangePassword {
     @Autowired
     private UserService UserService;
 
-    @PostMapping("/change-password")
-    public ResponseEntity<String> changePassword(@RequestBody ChangePasswordRequestDto changePasswordRequestdto) {
-        try {
-            UserService.changePassword(changePasswordRequestdto.getEmail(),
-                    changePasswordRequestdto.getOldPassword(),
-                    changePasswordRequestdto.getNewPassword());
-            return ResponseEntity.ok("Mật khẩu đã được thay đổi thành công");
-        } catch (RuntimeException ex) {
-            return ResponseEntity.badRequest().body(ex.getMessage());
-        }
-    }
+
 }
