+++ TuneBox/Customer/src/main/java/org/example/customer/API/UserController.java
@@ -1,14 +1,21 @@
 package org.example.customer.API;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
+import org.example.library.dto.ChangePasswordRequestDto;
+import org.example.library.dto.RequestSignUpModel;
 import org.example.library.dto.UserDto;
+import org.example.library.model.RespondModel;
 import org.example.library.repository.UserRepository;
 import org.example.library.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -16,7 +23,7 @@ import java.util.Map;
 
 @CrossOrigin(origins = "http://localhost:3000", allowCredentials = "true")
 @RestController
-@RequestMapping("/User")
+@RequestMapping("/user")
 public class UserController {
     @Autowired
     private UserService UserService;
@@ -24,12 +31,13 @@ public class UserController {
     @Autowired
     private UserRepository Repo;
 
-
+private BCryptPasswordEncoder passwordEncoder;
 
     //LOGIN
     @PostMapping("/log-in")
     public ResponseEntity<?> login(@RequestBody UserDto user) {
         Map<String, Object> response = new HashMap<>();
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
         try {
             UserDto loggedInUser = UserService.Login(user);
             response.put("status", true);
@@ -47,6 +55,7 @@ public class UserController {
     //FORGOT PASSWORD
     @PostMapping("/forgot-password")
     public ResponseEntity<String> forgotPassword(@RequestBody  UserDto user) {
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
         try {
             UserService.ForgotPassword(user);
             return ResponseEntity.ok("Email reset password đã được gửi đến địa chỉ của bạn");
@@ -58,9 +67,26 @@ public class UserController {
     // RESET PASSWORD
     @PostMapping("/reset-password")
     public ResponseEntity<?> resetPassword(@RequestBody UserDto user) {
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
         try {
             UserService.resetPassword(user.getToken(), user.getNewPassword());
             return ResponseEntity.ok("Mật khẩu đã được đặt lại thành công");
+        } catch (RuntimeException ex) {
+            return ResponseEntity.badRequest().body(ex.getMessage());
+        }
+    }
+//Change Password
+    @PostMapping("/change-password")
+    public ResponseEntity<String> changePassword(@RequestBody ChangePasswordRequestDto changePasswordRequestdto) {
+
+        //encode
+        changePasswordRequestdto.setOldPassword(passwordEncoder.encode(changePasswordRequestdto.getOldPassword()));
+        changePasswordRequestdto.setNewPassword(passwordEncoder.encode(changePasswordRequestdto.getNewPassword()));
+        try {
+            UserService.changePassword(changePasswordRequestdto.getEmail(),
+                    changePasswordRequestdto.getOldPassword(),
+                    changePasswordRequestdto.getNewPassword());
+            return ResponseEntity.ok("Mật khẩu đã được thay đổi thành công");
         } catch (RuntimeException ex) {
             return ResponseEntity.badRequest().body(ex.getMessage());
         }
@@ -80,35 +106,105 @@ public class UserController {
     }
 
 
-    @GetMapping("/login/oauth2/success")
-    public ResponseEntity<?> loginWithGoogle(Authentication authentication) {
+//    @GetMapping("/login/oauth2/success")
+//    public ResponseEntity<?> loginWithGoogle(Authentication authentication) {
+//        Map<String, Object> response = new HashMap<>();
+//
+//        if (authentication == null || !(authentication.getPrincipal() instanceof OAuth2User)) {
+//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Authentication is required");
+//        }
+//
+//        OAuth2User oauth2User = (OAuth2User) authentication.getPrincipal();
+//
+//        // Kiểm tra oauth2User có phải null hay không
+//        if (oauth2User == null) {
+//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("OAuth2 User is not found");
+//        }
+//
+//        String email = oauth2User.getAttribute("email");
+//        String name = oauth2User.getAttribute("name");
+//
+//        if (email == null || name == null) {
+//            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Email or name not found");
+//        }
+//
+//        UserDto userDto = UserService.loginWithGoogle(email, name);
+//
+//        response.put("status", true);
+//        response.put("message", "Đăng nhập thành công");
+//        response.put("data", userDto);
+//
+//        return ResponseEntity.ok(response);
+//    }
+
+    @PostMapping("/sign-up")
+    public ResponseEntity<?> Register(@RequestBody RequestSignUpModel requestSignUpModel,
+                                      HttpServletRequest request
+    ) {
+        String psEncode =passwordEncoder.encode(requestSignUpModel.getUserDto().getPassword());
+        requestSignUpModel.getUserDto().setPassword(psEncode);
+        RespondModel response = new RespondModel();
+        try {
+            response.setMessage("Đăng ký thành công");
+            response.setData(UserService.Register(requestSignUpModel));
+            response.setStatus(true);
+            HttpSession session = request.getSession(true);
+            session.setAttribute("userId", requestSignUpModel.getUserDto().getId());
+        } catch (Exception ex) {
+            response.setMessage(ex.getMessage());
+            response.setData(null);
+            response.setStatus(false);
+        }
+        return ResponseEntity.ok(response);
+    }
+    @GetMapping("/get/{id}")
+    public ResponseEntity<?> GetUser(@PathVariable("id") Long UserId){
         Map<String, Object> response = new HashMap<>();
 
-        if (authentication == null || !(authentication.getPrincipal() instanceof OAuth2User)) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Authentication is required");
+        try {
+            response.put("status", true);
+            response.put("message", "Succesfull");
+            response.put("data", UserService.findById(UserId));
+        } catch (Exception ex) {
+            response.put("status", false);
+            response.put("message", "Fail");
+            response.put("data", null);
         }
-
-        OAuth2User oauth2User = (OAuth2User) authentication.getPrincipal();
-
-        // Kiểm tra oauth2User có phải null hay không
-        if (oauth2User == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("OAuth2 User is not found");
-        }
-
-        String email = oauth2User.getAttribute("email");
-        String name = oauth2User.getAttribute("name");
-
-        if (email == null || name == null) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Email or name not found");
-        }
-
-        UserDto userDto = UserService.loginWithGoogle(email, name);
-
-        response.put("status", true);
-        response.put("message", "Đăng nhập thành công");
-        response.put("data", userDto);
 
         return ResponseEntity.ok(response);
+    }
+
+    @GetMapping("/session")
+    public ResponseEntity<?> getSessionData(HttpServletRequest request) {
+        Map<String, Object> response = new HashMap<>();
+
+        HttpSession session = request.getSession(false);
+        if (session != null) {
+            Object userId = session.getAttribute("userId");
+            if (userId != null) {
+                response.put("status", true);
+                response.put("message", "Dữ liệu session được tìm thấy");
+                response.put("data", userId);
+            } else {
+                response.put("status", false);
+                response.put("message", "Không tìm thấy user ID trong session");
+            }
+        } else {
+            response.put("status", false);
+            response.put("message", "Session không tồn tại");
+        }
+
+        return ResponseEntity.ok(response);
+    }
+
+    @PostMapping("/check")
+    public ResponseEntity<String> check(@RequestBody RequestSignUpModel requestSignUpModel) {
+        try {
+            UserService.CheckLogin(requestSignUpModel);
+            return ResponseEntity.ok("Check completed successfully");
+        } catch (ResponseStatusException ex) {
+            return ResponseEntity.status(ex.getStatusCode()).body(ex.getReason()); // Use getStatusCode()
+        }
     }
 
 }
