+++ TuneBox/Customer/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.application.name=Customer
-server.port=8081
+server.port=8080
 
 
 #Database
@@ -18,4 +18,12 @@ spring.servlet.multipart.enabled= true
 spring.servlet.multipart.file-size-threshold = 2KB
 spring.servlet.multipart.max-file-size = 100MB
 spring.servlet.multipart.max-request-size= 100MB
+
+#mail
+spring.mail.host=smtp.gmail.com
+spring.mail.port=587
+spring.mail.username=phongphps32337@fpt.edu.vn
+spring.mail.password=khlo flkn uidk rtom
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.starttls.enable=true
