+++ TuneBox/Library/src/main/java/org/example/library/service/UserService.java
@@ -17,7 +17,7 @@ public interface UserService {
 
     void resetPassword(String token, String newPassword);
 
-    UserDto loginWithGoogle(String email , String name);
+//    UserDto loginWithGoogle(String email , String name);
     UserDto Login(UserDto user);
     void changePassword(String email, String oldPassword, String newPassword);
 }
